reorganize genomicsdb components
