+++ apps/genomicsdb/app/user/profile/page.tsx
@@ -1,0 +1,5 @@
+const UserProfilePage = () => {
+    return <div></div>;
+}
+
+export default UserProfilePage;
